Get rid of transitive compile-time dependencies in the resources macro
