+++ helix/lib/game/process/resources/ram.ex
@@ -1,3 +1,3 @@
-defmodule Game.Process.Resources.CPU do
+defmodule Game.Process.Resources.RAM do
   use Game.Process.Resources.Behaviour.Default
 end
